pyproject.toml
[project]
name = "KDSautolib"
version = "0.0.4"
authors = [
  { name="kmhswimgirl", email="kmhswimgirl@gmail.com" },
]
description = "A python wrapper for automating KDS syringe pump devices"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = ["pyserial>=3.5"]
license = "MIT"
license-files = ["LICEN[CS]E*"]

[project.urls]
Homepage = "https://github.com/kmhswimgirl/kds-lib"
Issues = "https://github.com/kmhswimgirl/kds-lib/issues"

[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"
